Fix method prototype for NSInteger/CGFloat usage
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40319 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSTableHeaderView.m
@@ -54,8 +54,8 @@
 #define mouse_sensitivity 4
 
 @interface NSTableView (GNUstepPrivate)
-- (void) _userResizedTableColumn: (int)index
-                           width: (float)width;
+- (void) _userResizedTableColumn: (NSInteger)index
+                           width: (CGFloat)width;
 - (CGFloat *) _columnOrigins;
 - (void) _mouseDownInHeaderOfTableColumn: (NSTableColumn *)tc;
 - (void) _clickTableColumn: (NSTableColumn *)tc;
